requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tortoise-orm==0.20.0
aiomysql==0.2.0
aerich==0.7.2
python-dotenv==0.19.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
wfdb==4.1.2
